Stop bubble bar from blinking post-slide
CreateQuestionContent lived in template.tsx, which Next.js App Router
re-instantiates on every navigation. Each route change unmounted the
component, cleared the draft-poll-portal state, and remounted it; the
brief gap between unmount and the next MutationObserver rAF showed up
as visible blinks of the category bubble bar after a slide-overlay
transition (often 2-4 times in dev with StrictMode double-mounts).

Same mechanism that forces SlideOverlayHost to live in the root
layout. Move CreateQuestionContent there too via a thin client wrapper
PersistentCreatePollHost — the component mounts once per page load and
stays mounted across client-side navigation, so the portal state and
its MutationObserver subscription are continuous.

PersistentCreatePollHost also takes over the create-poll chunk's idle
preload (previously in template.tsx, where it re-scheduled on every
navigation).
