+++ CLAUDE.md
@@ -2031,7 +2031,11 @@ bash scripts/remote.sh "docker exec whoeverwants-db-1 psql -U whoeverwants -c \"
 - **No bottom bar. No home button.** The old three-button bottom bar (Home / + / Profile) was removed. Navigation is:
   - **Floating "+" FAB on home only**: a single circular blue "+" button pinned bottom-right via `position: fixed` + `max(1.5rem, env(safe-area-inset-right, 0px))` / `max(1rem, env(safe-area-inset-bottom, 0px))`. Tapping it navigates to `/p/` (the empty placeholder), where the draft poll card's category bubble bar lets the user pick a category to start a new poll. Home does NOT show the bubble bar — choosing a category is a per-group decision, not a "starting fresh" decision.
   - **Category bubble bar on group-like pages**: rendered IN-FLOW inside the always-on draft poll card on the group page, NOT as a floating bottom-fixed bar. One bubble per `BUILT_IN_TYPES` entry plus an "Other" custom bubble — see `BUBBLE_ENTRIES` in `app/create-poll/page.tsx`. Each bubble calls `openModalFor(category)` which seeds a fresh draft via `emptyDraft({category})` and opens the new-question modal. (Earlier iterations had a floating What/When/Where 3-bubble bar with `?mode=time` / `?category=restaurant` URL preselection; that's gone — see the historical note in the Active Plan section above.)
-  - **`LazyCreateQuestionContent` mounts unconditionally on `isMounted`, NOT gated on `isGroupLikePage`.** The component owns the bubble-bar portal AND the create-poll modal. Earlier it was gated on `isGroupLikePage` so it only mounted after pathname flipped to a group route — which meant the overlay-slide transition (`slideToGroup`) showed an empty `#draft-poll-portal` for the full duration of the slide, and the bubble bar popped into view ~150-300ms AFTER the slide visually completed. Symptom: "buttons at the bottom don't appear until after the page finishes sliding." Mounting unconditionally lets the always-on `MutationObserver` inside `CreateQuestionContent` detect the overlay's `#draft-poll-portal` the moment the overlay enters the DOM, well before `router.push` flips the pathname. The component renders nothing visible when there's no portal target AND no open modal, so mounting it on the home page costs nothing visual — only its lazy module's effects (form-state load from localStorage, `getUserName` lookup, MutationObserver attach). The lazy module is already preloaded via the existing `requestIdleCallback` preload, so the mount cost is one React commit. Don't re-introduce a route-based mount gate without an alternative path for the slide-overlay case.
+  - **`LazyCreateQuestionContent` lives in `<PersistentCreatePollHost />` inside `app/layout.tsx`, NOT `app/template.tsx`.** The component owns the bubble-bar portal AND the create-poll modal. Two previous gating patterns failed:
+    1. **Route-gated** (`isMounted && isGroupLikePage`): the component only mounted after `router.push` flipped the pathname, so the overlay-slide transition showed an empty `#draft-poll-portal` for the full slide duration; bubble bar popped in ~150-300ms after the slide visually completed.
+    2. **Unconditional in template.tsx** (`isMounted` only): per Next.js App Router semantics, `template.tsx` re-instantiates on every navigation — so the component unmounted + remounted across home→group, and the brief state-init→MutationObserver-fire gap caused the bubble bar to BLINK off and back on once per navigation tick (often 2-4 times in StrictMode dev). Same mechanism that forces `<SlideOverlayHost />` to live in the layout.
+
+    Layout-level mounting persists across client-side navigation — the component mounts once per page load and stays mounted. The `<PersistentCreatePollHost />` wrapper handles the lazy import + chunk-error reload + idle preload internally (the preload was previously in template.tsx, where it re-ran on every navigation). The component renders nothing visible when there's no portal target AND no open modal, so mounting on every route (including home) costs nothing visual — only its lazy module's effects (form-state load from localStorage, `getUserName` lookup, MutationObserver attach). The idle preload runs once per page load. Don't try to move the mount back into `template.tsx` without solving the re-instantiation flicker.
   - **`CreateQuestionContent`'s portal lookup uses `querySelectorAll` and picks the LAST match, not `getElementById`.** During the slide there are TWO `#draft-poll-portal` divs in the DOM simultaneously: the real route's (inside the React tree under `#__next`) and the overlay's (createPortal'd directly to `document.body`, so appended LATER in DOM order). `getElementById` returns the FIRST match — the real route's — which is hidden behind the overlay's `z-index: 60` layer, so the bubble bar would be portaled into the wrong (invisible) target. The "last-in-DOM" rule picks the overlay's portal during the slide and falls back seamlessly to the real route's once the overlay unmounts (their visual positions coincide, so the user sees no flicker). Verified end-to-end via DOM trace: at t=200ms the bubble bar is in the overlay's portal with URL still `/`; at t=500ms URL has flipped, two portals exist, bubble bar stays in the overlay; at t=1000ms+ overlay is gone and the bubble bar has hopped to the real-route portal. Don't switch back to `getElementById` "for simplicity" — the dual-portal coexistence is load-bearing for the slide animation.
   - **Settings gear**: only on the home page, upper-left, icon-only (no text). Links to `/settings`. Rendered as `position: absolute` inside a `relative` wrapper around just the h1, with `top-1/2 -translate-y-1/2` so its vertical center auto-tracks the title's midline (no hardcoded offset — survives font-size/padding changes). Sits in normal page flow and **scrolls off-screen with the page** (intentionally not fixed). The outer container's `padding-top` (`calc(0.75rem + env(safe-area-inset-top, 0px))`) handles the iOS notch clearance.
   - **Back arrow**: the HeaderPortal back button only renders on the settings page when there's in-app history; all other pages (group, question) render their own back button in their fixed header.
